+++ lib/attributor/extensions/randexp.rb
@@ -7,4 +7,9 @@ class Randgen
   def self.date
     return DATE_TIME_EPOCH - rand(800)
   end
+
+  def self.time
+    return date.to_time
+  end
+
 end