+++ CHANGELOG.md
@@ -7,6 +7,7 @@ next
 * Added `recurse` option to `Type.load` that is used by `Model` and `Hash` to force the loading of values (specifically, so that default values are assigned) even if the loaded value is `nil`.
 * Fix `Attributor::CSV` to dump `String` values and generate `String` examples.
 * Default values of `false` now work correctly.
+* Added `BigDecimal`, `Date` and `Time` types
 
 2.2.1
 ------